REF: Reorganize benchmarks so they run as a module
