+++ benchmarking/__main__.py
@@ -50,6 +50,8 @@
 PyCTest driver for TomoPy
 """
 
+from __future__ import absolute_import
+
 import os
 import sys
 import glob
